// packages/@shared/models/lib/managers/_types.ts
export const enum CryptoActionNames {
    SET_SECRET = "CRYPTO_SET_SECRET",
    RETRIEVE_SECRET = "CRYPTO_RETRIEVE_SECRET",
    IS_SECRET_INITIALIZED = "IS_SECRET_INITIALIZED",
    RESET = "CRYPTO_RESET",
}

export const enum AccountActionNames {
    SET_PASSWORD = "ACC_SET_PASSWORD",
    RETRIEVE_PASSOWRD = "ACC_RETRIEVE_PASSOWRD",
    SET_AUTHORIZED = "ACC_SET_AUTHORIZED",
    RETRIEVE_AUTHORIZED = "ACC_RETRIEVE_AUTHORIZED",
    RESET = "ACC_RESET",
}